V11/remove old guard rails (#1847)
* Delete a portion of the RSpec double verification that is not auto enabled

* Add in a bit of a split for some of the glue code
