+++ lib/cucumber/glue/nil_world.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+module Cucumber
+  module Glue
+    # Raised if a World block returns Nil.
+    class NilWorld < StandardError
+      def initialize
+        super('World procs should never return nil')
+      end
+    end
+  end
+end
