+++ lib/cucumber/rspec/doubles.rb
@@ -5,15 +5,9 @@ require 'rspec/mocks'
 World(RSpec::Mocks::ExampleMethods)
 
 Before do
-  if RSpec::Mocks::Version::STRING >= '2.9.9'
-    RSpec::Mocks.setup
-  else
-    RSpec::Mocks.setup(self)
-  end
+  RSpec::Mocks.setup(self)
 end
 
 After do
-  RSpec::Mocks.verify
-ensure
   RSpec::Mocks.teardown
 end
